move towards 3 tabs
