+++ src/stelnet/board/query/provider/DmodProvider.java
@@ -1,0 +1,24 @@
+package stelnet.board.query.provider;
+
+import com.fs.starfarer.api.impl.campaign.DModManager;
+import com.fs.starfarer.api.impl.campaign.ids.Tags;
+import com.fs.starfarer.api.loading.HullModSpecAPI;
+import java.util.Collections;
+import java.util.List;
+
+public class DmodProvider {
+
+    private static transient List<HullModSpecAPI> allDMods;
+
+    public static void reset() {
+        allDMods = null;
+    }
+
+    public List<HullModSpecAPI> getDMods() {
+        if (allDMods == null) {
+            allDMods = DModManager.getModsWithTags(Tags.HULLMOD_DAMAGE);
+            Collections.sort(allDMods, new ShipHullSpecSorter());
+        }
+        return allDMods;
+    }
+}
