+++ src/stelnet/board/query/view/add/ButtonGroup.java
@@ -5,6 +5,7 @@ import com.fs.starfarer.api.util.Misc;
 import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
+import stelnet.board.query.view.FilteringButton;
 import stelnet.util.ColorUtils;
 import stelnet.util.L10n;
 import uilib.Button;
