+++ src/stelnet/board/query/view/add/SelectDeselectButton.java
@@ -4,6 +4,7 @@ import com.fs.starfarer.api.ui.CutStyle;
 import com.fs.starfarer.api.ui.IntelUIAPI;
 import com.fs.starfarer.api.ui.TooltipMakerAPI;
 import com.fs.starfarer.api.util.Misc;
+import stelnet.board.query.view.FilteringButton;
 import stelnet.util.L10n;
 import uilib.Button;
 import uilib.EventHandler;
