+++ src/stelnet/board/query/view/add/ShipButtonUtils.java
@@ -8,8 +8,7 @@ import java.util.LinkedList;
 import java.util.List;
 import stelnet.CommonL10n;
 import stelnet.board.query.provider.ShipProvider;
-import stelnet.filter.FleetMemberHasDMod;
-import stelnet.filter.FleetMemberHasDModCount;
+import stelnet.board.query.view.FilteringButton;
 import stelnet.filter.ShipHullHasBays;
 import stelnet.filter.ShipHullHasBuiltIn;
 import stelnet.filter.ShipHullIsManufacturer;
@@ -83,24 +82,5 @@ public class ShipButtonUtils {
             );
         }
         return builtInList.toArray(new FilteringButton[] {});
-    }
-
-    public static FilteringButton[] getDMods(ShipProvider provider) {
-        List<FilteringButton> dMods = new LinkedList<>();
-        for (HullModSpecAPI dMod : provider.getDMods()) {
-            dMods.add(
-                new FilteringButton(dMod.getDisplayName(), new FleetMemberHasDMod(dMod.getId(), dMod.getDisplayName()))
-            );
-        }
-        return dMods.toArray(new FilteringButton[] {});
-    }
-
-    public static FilteringButton[] getDModsCount() {
-        List<FilteringButton> dModCount = new LinkedList<>();
-        dModCount.add(new FilteringButton(CommonL10n.NONE, new FleetMemberHasDModCount(0)));
-        for (int i = 1; i < 6; i++) {
-            dModCount.add(new FilteringButton(String.valueOf(i), new FleetMemberHasDModCount(i)));
-        }
-        return dModCount.toArray(new FilteringButton[] {});
     }
 }
