+++ src/stelnet/board/query/view/add/ShipQueryFactory.java
@@ -10,11 +10,10 @@ import lombok.Getter;
 import stelnet.CommonL10n;
 import stelnet.board.query.QueryL10n;
 import stelnet.board.query.provider.ShipProvider;
+import stelnet.board.query.view.FilteringButton;
 import stelnet.board.query.view.dialog.PickerDialog;
 import stelnet.board.query.view.dialog.ShipPickerDialog;
 import stelnet.filter.Filter;
-import stelnet.filter.LogicalAnd;
-import stelnet.filter.LogicalOr;
 import stelnet.util.L10n;
 import uilib.Button;
 import uilib.Renderable;
@@ -23,7 +22,6 @@ import uilib.Renderable;
 public class ShipQueryFactory extends QueryFactory {
 
     private transient ShipProvider provider = new ShipProvider();
-    private transient ShipButtonHelper helper = new ShipButtonHelper(this);
 
     private final FilteringButton[] classSizes = ShipButtonUtils.getClassSizes();
     private final FilteringButton[] mountSizes = ShipButtonUtils.getMountSizes();
@@ -31,24 +29,10 @@ public class ShipQueryFactory extends QueryFactory {
     private final FilteringButton[] mountBays = ShipButtonUtils.getMountBays(this);
     private final FilteringButton[] designTypes = ShipButtonUtils.getManufacturers(provider);
     private final FilteringButton[] builtIns = ShipButtonUtils.getBuiltIns(provider);
-    private final FilteringButton[] dModCount = ShipButtonUtils.getDModsCount();
-    private final FilteringButton[] dModAllowed = ShipButtonUtils.getDMods(provider);
 
     public Object readResolve() {
         provider = new ShipProvider();
-        helper = new ShipButtonHelper(this);
         return this;
-    }
-
-    public void addDmodFilters(Set<Filter> filters) {
-        addSelectedOrNone(filters, dModCount, L10n.get(QueryL10n.DMOD_COUNT), true);
-        Set<Filter> allowedDmods = getFilters(dModAllowed, true);
-        Set<Filter> disallowedDmods = getFilters(dModAllowed, false);
-        if (!hasDmodSelection()) {
-            return;
-        }
-        filters.add(new LogicalOr(allowedDmods, L10n.get(QueryL10n.DMOD_ALLOWED)));
-        filters.add(new LogicalAnd(negateFilters(disallowedDmods), L10n.get(QueryL10n.DMOD_DISALLOWED)));
     }
 
     public void setFighterBays(FighterBaysButton active) {
@@ -67,9 +51,6 @@ public class ShipQueryFactory extends QueryFactory {
     public Set<Filter> getFilters(boolean forResults) {
         Set<Filter> filters = getCommonFilters();
         addSelectedOrNone(filters, builtIns, L10n.get(QueryL10n.BUILT_IN), true);
-        if (forResults) {
-            addDmodFilters(filters);
-        }
         return filters;
     }
 
@@ -85,8 +66,7 @@ public class ShipQueryFactory extends QueryFactory {
 
     @Override
     protected List<Renderable> getQueryBuildingComponents() {
-        helper.prepareBuiltIns();
-        helper.prepareDmods();
+        prepareBuiltIns();
         List<Renderable> elements = new LinkedList<>();
         elements.add(new ButtonGroup(sizeHelper, QueryL10n.CLASS_SIZE, classSizes, true));
         elements.add(new SectionHeader(sizeHelper.getGroupAndTextWidth(), QueryL10n.WEAPON_MOUNTS, true));
@@ -95,9 +75,9 @@ public class ShipQueryFactory extends QueryFactory {
         elements.add(new ButtonGroup(sizeHelper, QueryL10n.FIGHTER_BAYS, mountBays, true));
         elements.add(new SectionHeader(sizeHelper.getGroupAndTextWidth(), QueryL10n.MANUFACTURERS, true, designTypes));
         elements.add(new ButtonGroup(sizeHelper, designTypes, true));
-        elements.add(new SectionHeader(sizeHelper.getGroupAndTextWidth(), QueryL10n.DMODS, true, dModAllowed));
-        elements.add(new ButtonGroup(sizeHelper, QueryL10n.DMOD_COUNT, dModCount, true));
-        elements.add(new ButtonGroup(sizeHelper, QueryL10n.DMOD_ALLOWED, dModAllowed));
+        // elements.add(new SectionHeader(sizeHelper.getGroupAndTextWidth(), QueryL10n.DMODS, true, dModAllowed));
+        // elements.add(new ButtonGroup(sizeHelper, QueryL10n.DMOD_COUNT, dModCount, true));
+        // elements.add(new ButtonGroup(sizeHelper, QueryL10n.DMOD_ALLOWED, dModAllowed));
         elements.add(new SectionHeader(sizeHelper.getGroupAndTextWidth(), QueryL10n.BUILT_IN, true, builtIns));
         elements.add(new ButtonGroup(sizeHelper, builtIns, true));
         return elements;
@@ -113,9 +93,12 @@ public class ShipQueryFactory extends QueryFactory {
         return filters;
     }
 
-    protected boolean hasDmodSelection() {
-        Set<Filter> allowedDmods = getFilters(dModAllowed, true);
-        Set<Filter> disallowedDmods = getFilters(dModAllowed, false);
-        return !allowedDmods.isEmpty() && !disallowedDmods.isEmpty();
+    public void prepareBuiltIns() {
+        Set<Filter> filters = getCommonFilters();
+        Set<String> hullModIds = getProvider().getBuiltInIds(filters);
+        hullModIds.add("None");
+        for (FilteringButton button : getBuiltIns()) {
+            button.updateVisibility(hullModIds);
+        }
     }
 }
