+++ src/stelnet/board/query/view/manage/SubmarketFilterButton.java
@@ -1,8 +1,9 @@
-package stelnet.board.query.view.list;
+package stelnet.board.query.view.manage;
 
 import com.fs.starfarer.api.ui.IntelUIAPI;
 import lombok.Setter;
 import stelnet.board.query.QueryManager;
+import stelnet.board.query.view.FilteringButton;
 import stelnet.filter.Filter;
 
 @Setter
